requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
psycopg2-binary==2.9.5
pytest==7.3.0
SQLAlchemy-Utils==0.41.1
Flask-Caching
flask-profiler